add : testStore, call in browser-node store.spec.ts
