+++ modules/utils/src/index.ts
@@ -26,3 +26,4 @@ export * from "./strings";
 export * from "./test";
 export * from "./transfers";
 export * from "./validateUpdateSignatures";
+export * from "./testStore";
